# frontend/openapi.json
{"openapi":"3.1.0","info":{"title":"Vocab","version":"0.1.0"},"paths":{"/api/paper":{"get":{"summary":"Get Paper","operationId":"get_paper_api_paper_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetTestPaperResponse"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/submit":{"post":{"summary":"Submit Paper","operationId":"submit_paper_api_submit_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TestPaper-Input"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostSubmitResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/result/specific":{"get":{"summary":"Get Result Of Paer","operationId":"get_result_of_paer_api_result_specific_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"paper_id","in":"query","required":true,"schema":{"type":"string","format":"uuid","title":"Paper Id"}},{"name":"test_id","in":"query","required":true,"schema":{"type":"string","format":"uuid","title":"Test Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetPaperResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/result/meta/me":{"get":{"summary":"Get My Result Only Meta","operationId":"get_my_result_only_meta_api_result_meta_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetResultResponse"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/result/meta":{"get":{"summary":"Get Student Result Only Meta","operationId":"get_student_result_only_meta_api_result_meta_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"student_id","in":"query","required":true,"schema":{"type":"string","format":"uuid","title":"Student Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetResultResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/students":{"get":{"summary":"Get Students","operationId":"get_students_api_students_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetStudentsResponse"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/sign_up":{"post":{"summary":"Sign Up","operationId":"sign_up_api_sign_up_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/sign_in":{"post":{"summary":"Sign In","operationId":"sign_in_api_sign_in_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_sign_in_api_sign_in_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/user/me":{"get":{"summary":"Get Me","operationId":"get_me_api_user_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}}},"components":{"schemas":{"APIStatus":{"type":"integer","enum":[20000,50001,50000],"title":"APIStatus"},"Answer":{"properties":{"u_id":{"type":"string","format":"uuid","title":"U Id"},"content":{"type":"string","title":"Content"},"checked":{"type":"boolean","title":"Checked","default":false}},"type":"object","required":["content"],"title":"Answer"},"Body_sign_in_api_sign_in_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"password"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_sign_in_api_sign_in_post"},"Candidate":{"properties":{"id":{"type":"integer","title":"Id","default":0},"u_id":{"type":"string","format":"uuid","title":"U Id"},"text":{"$ref":"#/components/schemas/Text"},"answer":{"type":"boolean","title":"Answer","default":false},"checked":{"type":"boolean","title":"Checked","default":false}},"type":"object","required":["text"],"title":"Candidate"},"Difficulty":{"type":"string","enum":["easy","moderate","hard"],"title":"Difficulty"},"GetPaperResponse":{"properties":{"request_id":{"type":"string","format":"uuid","title":"Request Id"},"status":{"$ref":"#/components/schemas/APIStatus","default":20000},"paper":{"$ref":"#/components/schemas/Paper"}},"type":"object","required":["paper"],"title":"GetPaperResponse"},"GetResultResponse":{"properties":{"request_id":{"type":"string","format":"uuid","title":"Request Id"},"status":{"$ref":"#/components/schemas/APIStatus","default":20000},"papers":{"items":{"$ref":"#/components/schemas/PaperMeta"},"type":"array","title":"Papers"}},"type":"object","required":["papers"],"title":"GetResultResponse"},"GetStudentsResponse":{"properties":{"request_id":{"type":"string","format":"uuid","title":"Request Id"},"status":{"$ref":"#/components/schemas/APIStatus","default":20000},"students":{"items":{"$ref":"#/components/schemas/UserDTO"},"type":"array","title":"Students"}},"type":"object","required":["students"],"title":"GetStudentsResponse"},"GetTestPaperResponse":{"properties":{"request_id":{"type":"string","format":"uuid","title":"Request Id"},"status":{"$ref":"#/components/schemas/APIStatus","default":20000},"paper":{"$ref":"#/components/schemas/TestPaper-Output"}},"type":"object","required":["paper"],"title":"GetTestPaperResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Paper":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"binded":{"$ref":"#/components/schemas/UserDTO"},"answer_map":{"type":"object","title":"Answer Map"},"problems":{"items":{"$ref":"#/components/schemas/Problem"},"type":"array","title":"Problems"}},"type":"object","required":["binded","answer_map","problems"],"title":"Paper"},"PaperMeta":{"properties":{"paper_id":{"type":"string","title":"Paper Id"},"test_id":{"type":"string","title":"Test Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["paper_id","test_id","created_at","updated_at"],"title":"PaperMeta"},"PostSubmitResponse":{"properties":{"request_id":{"type":"string","format":"uuid","title":"Request Id"},"status":{"$ref":"#/components/schemas/APIStatus","default":20000},"score":{"type":"integer","title":"Score"},"user":{"$ref":"#/components/schemas/UserDTO"}},"type":"object","required":["score","user"],"title":"PostSubmitResponse"},"Problem":{"properties":{"id":{"type":"integer","title":"Id"},"u_id":{"type":"string","format":"uuid","title":"U Id"},"difficulty":{"$ref":"#/components/schemas/Difficulty","default":"moderate"},"question_type":{"$ref":"#/components/schemas/QType","default":"korean"},"candidates":{"items":{"$ref":"#/components/schemas/Candidate"},"type":"array","title":"Candidates"}},"type":"object","required":["id","candidates"],"title":"Problem","description":"problem.question: Question()\n    - question.u_id는 problem.u_id과 같은 값\nproblem.answer: Answer()\n    - answer.u_id는 candidate.u_id와 같은 값"},"QA":{"properties":{"question":{"$ref":"#/components/schemas/Question"},"answers":{"items":{"$ref":"#/components/schemas/Answer"},"type":"array","title":"Answers"}},"type":"object","required":["question","answers"],"title":"QA"},"QType":{"type":"string","enum":["korean","english"],"title":"QType","description":"Korean -> 질문 한국어, 답변 영어"},"Question":{"properties":{"u_id":{"type":"string","format":"uuid","title":"U Id"},"content":{"type":"string","title":"Content"}},"type":"object","required":["content"],"title":"Question"},"Tag":{"type":"string","enum":["noun","pronoun","verb","adjective","adverb","preposition","conjunction","interjection","undecided"],"title":"Tag"},"TestPaper-Input":{"properties":{"paper_id":{"type":"string","format":"uuid","title":"Paper Id"},"test_id":{"type":"string","format":"uuid","title":"Test Id"},"binded":{"$ref":"#/components/schemas/UserDTO"},"q_a_set":{"items":{"$ref":"#/components/schemas/QA"},"type":"array","title":"Q A Set"}},"type":"object","required":["binded","q_a_set"],"title":"TestPaper"},"TestPaper-Output":{"properties":{"paper_id":{"type":"string","format":"uuid","title":"Paper Id"},"test_id":{"type":"string","format":"uuid","title":"Test Id"},"binded":{"$ref":"#/components/schemas/UserDTO"},"q_a_set":{"items":{"$ref":"#/components/schemas/QA"},"type":"array","title":"Q A Set"}},"type":"object","required":["binded","q_a_set"],"title":"TestPaper"},"Text":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"type":"string","title":"Name"},"tag":{"$ref":"#/components/schemas/Tag","default":"undecided"},"k_description":{"type":"string","title":"K Description"}},"type":"object","required":["id","name","k_description"],"title":"Text"},"Token":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"}},"type":"object","required":["access_token"],"title":"Token"},"UType":{"type":"string","enum":["student","teacher","guest","admin"],"title":"UType"},"UserCreate":{"properties":{"name":{"type":"string","title":"Name"},"password":{"type":"string","title":"Password"},"user_type":{"$ref":"#/components/schemas/UType","default":"student"}},"type":"object","required":["name","password"],"title":"UserCreate"},"UserDTO":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"name":{"type":"string","title":"Name","default":"admin"},"user_type":{"$ref":"#/components/schemas/UType","default":"student"}},"type":"object","title":"UserDTO"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"api/sign_in"}}}}}}